+++ lizard_fewsjdbc/management/commands/fews_jdbc_cache.py
@@ -1,0 +1,21 @@
+"""
+
+"""
+
+from django.core.management.base import BaseCommand, CommandError
+import logging
+
+from lizard_fewsjdbc.models import JdbcSource
+
+logger = logging.getLogger(__name__)
+
+
+class Command(BaseCommand):
+    args = ''
+    help = 'Populate fews jdbc cache for better user experience'
+
+    def handle(self, *args, **options):
+        for jdbc_source in JdbcSource.objects.all():
+            logger.info('Processing %s ...' % jdbc_source)
+            jdbc_source.get_filter_tree(ignore_cache=True)
+        logger.info('Finished successfully.')